QuestionPage : Extract Header and Sidebar to Layout
