+++ frontend/src/pages/QuestionPage.js
@@ -1,23 +1,10 @@
 import * as React from "react";
 import { useState } from "react";
-import { styled, createTheme, ThemeProvider } from "@mui/material/styles";
 import CssBaseline from "@mui/material/CssBaseline";
-import MuiDrawer from "@mui/material/Drawer";
 import Box from "@mui/material/Box";
-import MuiAppBar from "@mui/material/AppBar";
-import Toolbar from "@mui/material/Toolbar";
-import List from "@mui/material/List";
-import Typography from "@mui/material/Typography";
-import Divider from "@mui/material/Divider";
-import IconButton from "@mui/material/IconButton";
-import Badge from "@mui/material/Badge";
 import Container from "@mui/material/Container";
 import Grid from "@mui/material/Grid";
 import Paper from "@mui/material/Paper";
-import MenuIcon from "@mui/icons-material/Menu";
-import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import NotificationsIcon from "@mui/icons-material/Notifications";
-import NavBarList from "../components/NavBarList";
 import Question from "../components/question/Question";
 import ViewQuestion from "../components/question/ViewQuestion";
 import EditQuestion from "../components/question/EditQuestion";
@@ -25,59 +12,6 @@ import AddQuestion from "../components/question/AddQuestion";
 import { getQuestionsFromLocalStorage, getTokenFromLocalStorage, saveQuestionsToLocalStorage } from "../LocalStorage";
 import axios from 'axios';
 import useCookie from "../components/useCookie";
-
-const drawerWidth = 240;
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  zIndex: theme.zIndex.drawer + 1,
-  transition: theme.transitions.create(["width", "margin"], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(["width", "margin"], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
-
-const Drawer = styled(MuiDrawer, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-  "& .MuiDrawer-paper": {
-    position: "relative",
-    whiteSpace: "nowrap",
-    width: drawerWidth,
-    transition: theme.transitions.create("width", {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-    boxSizing: "border-box",
-    ...(!open && {
-      overflowX: "hidden",
-      transition: theme.transitions.create("width", {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.leavingScreen,
-      }),
-      width: theme.spacing(7),
-      [theme.breakpoints.up("sm")]: {
-        width: theme.spacing(9),
-      },
-    }),
-  },
-}));
-
-// TODO remove, this demo shouldn't need to reset the theme.
-const defaultTheme = createTheme({
-  palette: {
-    mode: "dark",
-  },
-});
 
 const COMPLEXITY = {
   EASY: "EASY",
@@ -139,10 +73,6 @@ function checkEmpty(
   return;
 }
 function QuestionPage() {
-    const [open, setOpen] = React.useState(true);
-    const toggleDrawer = () => {
-        setOpen(!open);
-    };
     const {getAuthCookie} = useCookie();
 
     const checkDuplicateTitle = (title, questions) => {
@@ -232,73 +162,8 @@ function QuestionPage() {
         fetchQuestions()
     }, [])
     return (
-        <ThemeProvider theme={defaultTheme}>
         <Box sx={{ display: "flex" }}>
             <CssBaseline />
-            <AppBar position="absolute" open={open}>
-            <Toolbar
-                sx={{
-                pr: "24px", // keep right padding when drawer closed
-                }}
-            >
-                <IconButton
-                edge="start"
-                color="inherit"
-                aria-label="open drawer"
-                onClick={toggleDrawer}
-                sx={{
-                    marginRight: "36px",
-                    ...(open && { display: "none" }),
-                }}
-                >
-                <MenuIcon />
-                </IconButton>
-                <Typography
-                component="h1"
-                variant="h6"
-                color="inherit"
-                noWrap
-                sx={{ flexGrow: 1 }}
-                >
-                QUESTIONS
-                </Typography>
-                <IconButton color="inherit">
-                <Badge badgeContent={4} color="secondary">
-                    <NotificationsIcon />
-                </Badge>
-                </IconButton>
-            </Toolbar>
-            </AppBar>
-            <Drawer variant="permanent" open={open}>
-            <Toolbar
-                sx={{
-                display: "flex",
-                alignItems: "center",
-                justifyContent: "flex-end",
-                px: [1],
-                }}
-            >
-                <IconButton onClick={toggleDrawer}>
-                <ChevronLeftIcon />
-                </IconButton>
-            </Toolbar>
-            <Divider />
-            <NavBarList/>
-            
-            </Drawer>
-            <Box
-            component="main"
-            sx={{
-                backgroundColor: (theme) =>
-                theme.palette.mode === "light"
-                    ? theme.palette.grey[100]
-                    : theme.palette.grey[900],
-                flexGrow: 1,
-                height: "100vh",
-                overflow: "auto",
-            }}
-            >
-            <Toolbar />
             <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
                 <Grid container spacing={3}>
                 {" "}
@@ -389,9 +254,8 @@ function QuestionPage() {
                 )}
                 </Grid>
             </Container>
-            </Box>
+
         </Box>
-        </ThemeProvider>
     );
 }
